Refactor CustomTabsConnectionHelper to InternalModule. Test methods that depends on it.
