+++ packages/expo-web-browser/android/src/main/java/expo/modules/webbrowser/CustomTabsConnectionHelper.java
@@ -1,115 +1,13 @@
 package expo.modules.webbrowser;
 
-import android.content.ComponentName;
-import android.content.Context;
 import android.net.Uri;
-import androidx.browser.customtabs.CustomTabsClient;
-import androidx.browser.customtabs.CustomTabsServiceConnection;
-import androidx.browser.customtabs.CustomTabsSession;
 
-import org.unimodules.core.interfaces.LifecycleEventListener;
+import org.unimodules.core.interfaces.InternalModule;
 
-public class CustomTabsConnectionHelper extends CustomTabsServiceConnection implements LifecycleEventListener {
+public interface CustomTabsConnectionHelper extends InternalModule {
+  void warmUp(String packageName);
 
-  private Context mContext;
-  private String mPackageName;
-  private DeferredClientActionsQueue<CustomTabsClient> clientActions = new DeferredClientActionsQueue<>();
-  private DeferredClientActionsQueue<CustomTabsSession> sessionActions = new DeferredClientActionsQueue<>();
+  void mayInitWithUrl(String packageName, Uri uri);
 
-  CustomTabsConnectionHelper(Context context) {
-    this.mContext = context;
-  }
-
-  void warmUp(String packageName) {
-    clientActions.executeOrQueueAction(client -> client.warmup(0));
-    ensureConnection(packageName);
-  }
-
-  void mayInitWithUrl(String packageName, Uri uri) {
-    sessionActions.executeOrQueueAction(session -> session.mayLaunchUrl(uri, null, null));
-    ensureConnection(packageName);
-    ensureSession();
-  }
-
-  private void ensureSession() {
-    if (!sessionActions.hasClient()) {
-      clientActions.executeOrQueueAction(
-          client ->
-              sessionActions.setClient(client.newSession(null)));
-    }
-  }
-
-  boolean coolDown(String packageName) {
-    if (packageName.equals(this.mPackageName)) {
-      unbindService();
-      return true;
-    }
-    return false;
-  }
-
-  private void ensureConnection(String packageName) {
-    if (this.mPackageName != null && !this.mPackageName.equals(packageName)) {
-      clearConnection();
-    }
-    if (!connectionStarted(packageName)) {
-      CustomTabsClient.bindCustomTabsService(getContext(), packageName, this);
-      this.mPackageName = packageName;
-    }
-  }
-
-  private boolean connectionStarted(String packageName) {
-    return packageName.equals(this.mPackageName);
-  }
-
-  @Override
-  public void onBindingDied(ComponentName componentName) {
-    if (componentName.getPackageName().equals(this.mPackageName)) {
-      clearConnection();
-    }
-  }
-
-  @Override
-  public void onCustomTabsServiceConnected(ComponentName componentName, CustomTabsClient client) {
-    if (componentName.getPackageName().equals(mPackageName)) {
-      clientActions.setClient(client);
-    }
-  }
-
-  @Override
-  public void onServiceDisconnected(ComponentName componentName) {
-    if (componentName.getPackageName().equals(this.mPackageName)) {
-      clearConnection();
-    }
-  }
-
-  @Override
-  public void onHostResume() {
-    // do nothing
-  }
-
-  @Override
-  public void onHostPause() {
-    // do nothing
-  }
-
-  @Override
-  public void onHostDestroy() {
-    unbindService();
-  }
-
-  private void unbindService() {
-    getContext().unbindService(this);
-    clearConnection();
-  }
-
-  private void clearConnection() {
-    this.mPackageName = null;
-    clientActions.clear();
-    sessionActions.clear();
-  }
-
-  private Context getContext() {
-    return mContext;
-  }
-
+  boolean coolDown(String packageName);
 }
