+++ packages/expo-web-browser/android/src/main/java/expo/modules/webbrowser/WebBrowserPackage.java
@@ -6,6 +6,7 @@ import org.unimodules.core.BasePackage;
 import org.unimodules.core.ExportedModule;
 import org.unimodules.core.interfaces.InternalModule;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
@@ -13,11 +14,14 @@ public class WebBrowserPackage extends BasePackage {
 
   @Override
   public List<InternalModule> createInternalModules(Context context) {
-    return Collections.singletonList(new InternalCustomTabsActivitiesHelper());
+    List<InternalModule> list = new ArrayList<>();
+    list.add(new InternalCustomTabsActivitiesHelper());
+    list.add(new InternalCustomTabsConnectionHelper(context));
+    return list;
   }
 
   @Override
   public List<ExportedModule> createExportedModules(Context context) {
-    return Collections.singletonList((ExportedModule) new WebBrowserModule(context));
+    return Collections.singletonList(new WebBrowserModule(context));
   }
 }
